+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 - Added ApiResultException class.
 - Added getJsonEncodeOptions method to ApiController class.
+- Throws exception if result could not be JSON-encoded.
 
 ## [2.0.0] - 2018-05-14
 - **Backward compatibility break**: Changed parameter order in constructor in ApiResult class.